refactor(board): boardService 도메인별 7개 모듈 분할 (Phase 3)
- board/: shared(상수·타입·헬퍼·권한가드·공통조회) + post/comment/reaction/report/schedule/substitute 서비스
- import 단방향(shared 최하위, post→schedule, comment→post)으로 순환 차단
- 진입점 boardService.ts(1808→88)는 7모듈 re-export + boardService 객체 + default 유지 → 소비자·테스트 무손상
- 본문·시그니처 불변, 코드 이동만 (tsc 0, test 4286 pass)
